backend/requirements.txt
fastapi[all]==0.111.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
boto3
pydantic-settings==2.3.1
uvicorn==0.30.1
httpx
email-validator